feat(aichat): agent add memory
